+++ packages/toolkit/src/main.ts
@@ -1,3 +1,4 @@
 export { Injector } from "@collidor/injector";
 export * from "@collidor/event";
 export * from "@collidor/command";
+export * from "@collidor/schema-command";